Adding a privacy considerations section
This is a non-final proposal to add a privacy considerations section to the specification to be clear about the slight fingerprinting risk in response to #36. This is not intended to be the final draft but a place to open discussion on the precise language to make it as clear as possible.